Codegen for naive GPU hash-join/hash-groupby, name-based look-ups for attributes in plans, codegen in pipelines for GPU-side, SSBM Q1.x plans and catalog info, dictionary for strings (GPU only)
Code generation for naive hash-join and hash-group-by using chains for
GPU. Attributes in plans can now simply be referenced as, eg,
"attributes": [{
	"relName": "inputs/ssbm/date.csv",
	"attrName": "d_datekey"
}]
and the parser will fetch the necessary information from the catalog.
Code generation for the GPU side now produces a vector of pipelines,
with each pipeline having an open-consume(block)-close interface.
Simple plans for the first group of SSBM queries and corresponding
catalog information are now included (input files are not included due
to the file size).
Initial support for string dictionaries for the GPU side.