First operators for the exchange and initial multigpu support
Option for setting the size of auxilary structures for GPU join and
group-by. PLan parsing is now happening depth-first to allow child
operators to define new RecordAttributes. Added an option in
binary-col-plugin to select whether the column size is extracted from
the contents of the files (first entry of each file) or from the file
size. Introduction of RawPipeline to drive the code generation of each
pipeline and moved part of the responsibilities of GpuRawContext to this
class. GpuRawContext now functions as a stack of RawPipelines.
Introduced a BlockType. Single producer Exchange operator, CpuToGpu
operators, HashRearrange, MemMove and a ScanInBlocks plugin. Added
multigpu engine as a submodule to be able to use parts of the existing
infrastructure.