Flush results to memory file, flushing for GPU-reduce, change optimizations to opt's O3, PipelineGen/LLVM-IR-to-MC clean up
Flush results to memory file to avoid the overhead (4ms overhead in a
16ms query for opening the output file! 0ms overhead for memory mapped
file). Flushing pipeline for GPU-reduce. Now opt's O3 optimizations are
applied to the generated code. Huge clean-up and redesign of Pipeline
Generators (RawPipelineGen) and LLVM IR to machine code related API.
Separated pipeline generator from IR-to-machine-code converters. Codegen
for init/deinit for the GPU side. The init/deinit code runs on the CPU
side. In addition, a CPU-side wrapper around the GPU consume is now
generated.