Make memset and memcpy's in open/close blocking for the current thread
Launching memset in the default stream does not guarantee that the host
will block neither that the memset will happen synchronously with other
streams. Only stream 0 and its associated streams will see the change.
This commit cahnges some memset's and memcpy to be blocking for the host
thread calling them, to maintain the abstraction.