Programmatic discovery of NVMe storage devices.
Currently, we have a nvmeStorage per nvme namespace. This currently makes sense because we only use a single namespace per drive. However, if we start using multiple namespaces we will need to figure out how to reason about pcie bandwidth shared between namespaces on the same physical drive.

Further, we use the order that we discover NVMe drives for the unique id. This is fine currently, as we don't currently rely on being able to identify the same drive across instances of proteus. If we do, we need to perhaps use the partition uid, or the drives PCIe address. If we need to uniquely identify drives across a cluster of proteus instances, then PCIe address wont work.

We discover nvme drives by parsing /dev/*, we considered using nvme-cli, but using nvme-cli is not portable. Older versions on ubuntu 16.04 and 18.04 have considerably less information, which requires parsing devfs anyways for additional info such as linkspeed.

Discovery of nvme devices is not neccessary for in-memory operation of proteus.