Fix topology on systems without virtual nvme devices
There are two ways that sysfs stores information for nvme drives. One where /sys/block/nvme0n1 links to a pci device, and the other where it links to a virtual nvme device. We need to parse these cases separately to get to the physical pcie device. For more context see:
https://lore.kernel.org/all/3c725e5deaabaaf145f48f2f6fcfdae9f6d41e2e.camel@suse.de/t/#r543f7392cfa1ba8adfe5447735b8da6170022295