[storage] Add virtual mmmap load and test
This functionality enables delayed loading of data. It enables the storage manager to return handles for data that has not been loaded to memory yet. We are abusing pointers a little bit, because we return pointers that will segfault if anyone tries to dereference them.

The original motivation was to support caching, and to use the pointers returned by a virtual mmap_file as cache keys. And then use these cache keys in mem-move to access the real data in the cache.