[tooling] fix commandline build
We were not explicitly linking in librt in libolap, which is required
for shared memory. This did not always cause an issue in CLion, but did
in builds done on the command line. I am not sure why. It may depend
on which generator CMake is using. Regardless, we should explicitly
be linking against it instead of relying on a transitive link.

https://man7.org/linux/man-pages/man3/shm_open.3.html
https://refspecs.linuxbase.org/LSB_3.0.0/LSB-PDA/LSB-PDA/librt.html#STD-223-67-TBL-LIBRT-SHA-INTS