[platform] Make failure to construct IB non-fatal
When running in docker, ibv_get_device_list will find IB devices,
but attempting to construct an ib device lead to a runtime exception.
We now catch the exception log a warning and move on because this is
non-fatal when we don't use IB. We will need to revisit this when we
want to get IB working in docker.

Related issue:
 https://gitlab.epfl.ch/DIAS/PROJECTS/caldera/proteus/-/issues/58