[tooling] add gosu to docker requirements
gosu is a little utility for setting gid/uid. This does add a
little bloat to the CI container which doesn't need interactivity,
but it does let us use the same container image for CI and
interactive builds. The alternative is to have two containers.
The existing pelago-build and a new pelago-dev which extends it.
The downside of this approach is having to manage double the
images.