[OLAP] fix affinitizer case for CPU NUMA with no GPUs
If we had a multi NUMA system and only some NUMAs have GPUs,
then GPUAffinitizer::getLocalCUIndex(ptr) would result in a divide by
zero error if ptr referred to memory on a NUMA node with no GPUs.

Now, we search for the closest CPU NUMA node with a GPU and return a
random GPU from that CPU NUMA node.

This assumes that CPU NUMA distance is roughly approximate to CPU NUMA
to GPU distance.